Adding contents of C2000Ware v5.05.00.00

+++ driverlib/f28004x/driverlib/i2c.c
@@ -6,7 +6,7 @@
 //
 //###########################################################################
 // $Copyright:
-// Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com/
+// Copyright (C) 2025 Texas Instruments Incorporated - http://www.ti.com/
 //
 // Redistribution and use in source and binary forms, with or without 
 // modification, are permitted provided that the following conditions 
@@ -256,7 +256,7 @@ I2C_getInterruptStatus(uint32_t base)
     //
     // Return only the status bits associated with interrupts.
     //
-    temp = (uint32_t)(HWREGH(base + I2C_O_STR) & I2C_STR_INTMASK);
+    temp = (uint32_t)HWREGH(base + I2C_O_STR) & (uint32_t)I2C_STR_INTMASK;
 
     //
     // Read FIFO interrupt flags.
