+++ driverlib/f28p65x/driverlib/dcsm.h
@@ -6,7 +6,7 @@
 //
 //#############################################################################
 // 
-// C2000Ware v5.04.00.00
+// C2000Ware v5.05.00.00
 //
 // Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
 //
@@ -940,7 +940,7 @@ DCSM_forcePerSemStatus(DCSM_PERSem grabType, uint16_t value)
     //
     // Writes into the Flash wrapper semaphore register value
     //
-    HWREG(DCSMCOMMON_BASE + DCSM_O_PERSEM1) = ((value & 0x3U) <<
+    HWREG(DCSMCOMMON_BASE + DCSM_O_PERSEM1) = (((uint32_t)value & 0x3U) <<
                                                (uint32_t)grabType)|
                                               ((uint32_t)FLSEM_KEY <<
                                                DCSM_PERSEM1_KEY_S);
